Changes made to Food Class

+++ src/main/java/cube/storage/StorageManager.java
@@ -31,7 +31,7 @@ public class StorageManager {
 
 	}
 
-	public void storeRevenue(long revenue) {
+	public void storeRevenue(double revenue) {
 		
 	}
 }
